+++ src/core/engine.ts
@@ -26,7 +26,8 @@ import { createGame } from "../game/game";
 import { createCanvas } from "../gfx/canvas";
 import { initGfx } from "../gfx/gfx";
 import { initAppGfx } from "../gfx/gfxApp";
-import type { KAPLAYCtx, KAPLAYOpt } from "../types";
+import type { KAPLAYOpt } from "../types";
+import type { KAPLAYCtx } from "./contextType";
 import { startEngineLoop } from "./engineLoop";
 import { createFontCache } from "./fontCache";
 import { createFrameRenderer } from "./frameRendering";
