+++ src/ecs/components/misc/textInput.ts
@@ -1,6 +1,6 @@
 import type { KEventController } from "../../../events/events";
 import { _k } from "../../../shared";
-import type { Comp, GameObj, KAPLAYCtx } from "../../../types";
+import type { Comp, GameObj } from "../../../types";
 import type { TextComp } from "../draw/text";
 
 /**
@@ -51,7 +51,6 @@ export interface TextInputComp extends Comp {
 }
 
 export function textInput(
-    this: KAPLAYCtx,
     hasFocus: boolean = true,
     maxInputLength?: number,
 ): TextInputComp {
